requirements.txt
streamlit==1.35.0
pymongo==4.7.2
bcrypt==4.1.3
python-dotenv==1.0.1
google-generativeai==0.5.4
langchain==0.2.5
langchain-community==0.2.5
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pandas==2.2.2
openpyxl==3.1.4
PyPDF2==3.0.1
pdfplumber==0.11.1
tiktoken==0.7.0